admin panel button added

+++ main_service/models/user_service_models.py
@@ -32,7 +32,7 @@ class GetMovieRatingResponse(BaseModel):
 class Movie(BaseModel):
     movie_id: Optional[int] = None
     movie_title: Optional[str] = None
-    rating: Optional[int] = None
+    rating: Optional[float] = None
 
 
 class UserInfoRequest(BaseModel):
